pyproject.toml
[tool.poetry]
name = "myportfolio"
version = "0.1.0"
description = ""
authors = ["Praful Patekar <prafulpatekar23@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "src" }]


[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
whitenoise = "^6.8.2"
psycopg2 = "^2.9.10"
gunicorn = "^23.0.0"
pillow = "^11.0.0"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
